Convert SwitchWalletScreen to normal and modal
